==> 1. Crear base de datos.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import statsmodels.api as sm
import statsmodels.formula.api as smf
import pandas as pd

# Primera base de datos de Saber 11 2021 (principalmente Calendario B)

df1=pd.read_csv("Datos/SB11_20211.txt", sep = "¬" , encoding = 'utf-8', 
                   names=["ESTU_TIPODOCUMENTO", "ESTU_NACIONALIDAD", "ESTU_GENERO", "ESTU_FECHANACIMIENTO",
              "PERIODO","ESTU_CONSECUTIVO", "ESTU_ESTUDIANTE", "ESTU_PAIS_RESIDE", 
              "ESTU_TIENEETNIA", "ESTU_DEPTO_RESIDE", "ESTU_COD_RESIDE_DEPTO",
              "ESTU_MCPIO_RESIDE", "ESTU_COD_RESIDE_MCPIO", "FAMI_ESTRATOVIVIENDA",
              "FAMI_PERSONASHOGAR", "FAMI_CUARTOSHOGAR", "FAMI_EDUCACIONPADRE",
              "FAMI_EDUCACIONMADRE", "FAMI_TRABAJOLABORPADRE", "FAMI_TRABAJOLABORMADRE", 
              "FAMI_TIENEINTERNET", "FAMI_TIENECOMPUTADOR", "FAMI_TIENELAVADORA",
              "FAMI_TIENEHORNOMICROOGAS", "FAMI_TIENESERVICIOTV", "FAMI_TIENEAUTOMOVIL",
              "FAMI_TIENEMOTOCICLETA","FAMI_TIENECONSOLAVIDEOJUEGOS", "FAMI_NUMLIBROS",
              "FAMI_COMELECHEDERIVADOS","FAMI_COMECARNEPESCADOHUEVO", "FAMI_COMECEREALFRUTOSLEGUMBRE", 
              "FAMI_SITUACIONECONOMICA", "ESTU_DEDICACIONLECTURADIARIA","ESTU_DEDICACIONINTERNET",
              "ESTU_HORASSEMANATRABAJA", "ESTU_TIPOREMUNERACION", "COLE_CODIGO_ICFES",
              "COLE_COD_DANE_ESTABLECIMIENTO", "COLE_NOMBRE_ESTABLECIMIENTO", "COLE_GENERO",
                          "COLE_NATURALEZA","COLE_CALENDARIO", "COLE_BILINGUE", "COLE_CARACTER","COLE_COD_DANE_SEDE",
              "COLE_NOMBRE_SEDE", "COLE_SEDE_PRINCIPAL","COLE_AREA_UBICACION", "COLE_JORNADA",
              "COLE_COD_MCPIO_UBICACION", "COLE_MCPIO_UBICACION","COLE_COD_DEPTO_UBICACION",
              "COLE_DEPTO_UBICACION","ESTU_PRIVADO_LIBERTAD","ESTU_COD_MCPIO_PRESENTACION",
              "ESTU_MCPIO_PRESENTACION","ESTU_DEPTO_PRESENTACION","ESTU_COD_DEPTO_PRESENTACION",
              "PUNT_LECTURA_CRITICA","PERCENTIL_LECTURA_CRITICA","DESEMP_LECTURA_CRITICA",
              "PUNT_MATEMATICAS", "PERCENTIL_MATEMATICAS", "DESEMP_MATEMATICAS", "PUNT_C_NATURALES",
              "PERCENTIL_C_NATURALES","DESEMP_C_NATURALES", "PUNT_SOCIALES_CIUDADANAS",
              "PERCENTIL_SOCIALES_CIUDADANAS", "DESEMP_SOCIALES_CIUDADANAS",
              "PUNT_INGLES", "PERCENTIL_INGLES", "DESEMP_INGLES", "PUNT_GLOBAL",
              "PERCENTIL_ESPECIAL_GLOBAL", "PERCENTIL_GLOBAL","ESTU_ESTADOINVESTIGACION"])

df1.head()

df1 = df1.iloc[1:, :]

df1.head()

df1.shape

# Segunda base de datos de Saber 11 2021 (principalmente Calendario A)

df2=pd.read_csv("Datos/SB11_20212.txt", sep = "¬" , header=None, encoding = 'utf-8', 
                   names=["ESTU_TIPODOCUMENTO", "ESTU_NACIONALIDAD", "ESTU_GENERO",
                          "ESTU_FECHANACIMIENTO", "PERIODO", "ESTU_CONSECUTIVO",
                          "ESTU_ESTUDIANTE", "ESTU_PAIS_RESIDE", "ESTU_TIENEETNIA",
                          "ESTU_DEPTO_RESIDE", "ESTU_COD_RESIDE_DEPTO", "ESTU_MCPIO_RESIDE",
                          "ESTU_COD_RESIDE_MCPIO", "FAMI_ESTRATOVIVIENDA", "FAMI_PERSONASHOGAR",
                          "FAMI_CUARTOSHOGAR", "FAMI_EDUCACIONPADRE", "FAMI_EDUCACIONMADRE",
                          "FAMI_TRABAJOLABORPADRE", "FAMI_TRABAJOLABORMADRE", "FAMI_TIENEINTERNET",
                          "FAMI_TIENESERVICIOTV", "FAMI_TIENECOMPUTADOR", "FAMI_TIENELAVADORA",
                          "FAMI_TIENEHORNOMICROOGAS", "FAMI_TIENEAUTOMOVIL", "FAMI_TIENEMOTOCICLETA",
                          "FAMI_TIENECONSOLAVIDEOJUEGOS", "FAMI_NUMLIBROS", "FAMI_COMELECHEDERIVADOS",
                          "FAMI_COMECARNEPESCADOHUEVO", "FAMI_COMECEREALFRUTOSLEGUMBRE", 
                          "FAMI_SITUACIONECONOMICA", "ESTU_DEDICACIONLECTURADIARIA",
                          "ESTU_DEDICACIONINTERNET", "ESTU_HORASSEMANATRABAJA", "ESTU_TIPOREMUNERACION",
                          "COLE_CODIGO_ICFES", "COLE_COD_DANE_ESTABLECIMIENTO", "COLE_NOMBRE_ESTABLECIMIENTO",
                          "COLE_GENERO", "COLE_NATURALEZA", "COLE_CALENDARIO", "COLE_BILINGUE",
                          "COLE_CARACTER", "COLE_COD_DANE_SEDE", "COLE_NOMBRE_SEDE", "COLE_SEDE_PRINCIPAL",
                          "COLE_AREA_UBICACION", "COLE_JORNADA", "COLE_COD_MCPIO_UBICACION", "COLE_MCPIO_UBICACION",
                          "COLE_COD_DEPTO_UBICACION", "COLE_DEPTO_UBICACION", "ESTU_PRIVADO_LIBERTAD",
                          "ESTU_COD_MCPIO_PRESENTACION", "ESTU_MCPIO_PRESENTACION", "ESTU_DEPTO_PRESENTACION",
                          "ESTU_COD_DEPTO_PRESENTACION", "PUNT_LECTURA_CRITICA", "PERCENTIL_LECTURA_CRITICA", 
                          "DESEMP_LECTURA_CRITICA", "PUNT_MATEMATICAS", "PERCENTIL_MATEMATICAS", "DESEMP_MATEMATICAS",
                          "PUNT_C_NATURALES", "PERCENTIL_C_NATURALES", "DESEMP_C_NATURALES", "PUNT_SOCIALES_CIUDADANAS",
                          "PERCENTIL_SOCIALES_CIUDADANAS", "DESEMP_SOCIALES_CIUDADANAS", "PUNT_INGLES", 
                          "PERCENTIL_INGLES", "DESEMP_INGLES", "PUNT_GLOBAL", "PERCENTIL_GLOBAL", 
                          "PERCENTIL_ESPECIAL_GLOBAL", "ESTU_INSE_INDIVIDUAL", "ESTU_NSE_INDIVIDUAL", 
                          "ESTU_NSE_ESTABLECIMIENTO", "ESTU_ESTADOINVESTIGACION", "ESTU_GENERACION-E"
                       ])

df2.head()

df2 = df2.iloc[1:, :]

df2.head()

df2.shape

# Append las dos bases de datos

df = df1.append(df2, ignore_index = True)

df

# Exportar base de datos
df.to_csv("Saber2021.csv", index = False)
