==> 2. Depuración de los datos.ipynb <==
import pandas as pd

#leer grande base de datos
data = pd.read_csv("Datos/Saber2021.csv")

# Comprobar que subiera bien
data.head()

# Cuántas filas y columns
data.shape

# Ver los nombres de las columnas
data.columns

# Cuántos colegios, municipios y departamentos son representados
count = data[['COLE_COD_DANE_ESTABLECIMIENTO','COLE_COD_MCPIO_UBICACION', 
              'COLE_COD_DEPTO_UBICACION' ]].nunique()
print(count)

# Reducir el tamaño de la base, solo guardar variables más relevantes a la investigación
data = data.loc[: , ['ESTU_GENERO', 'PERIODO','ESTU_TIENEETNIA','FAMI_ESTRATOVIVIENDA',
                     'FAMI_TIENEINTERNET', 'FAMI_TIENECOMPUTADOR','FAMI_SITUACIONECONOMICA',
       'ESTU_DEDICACIONLECTURADIARIA', 'ESTU_DEDICACIONINTERNET',
       'ESTU_HORASSEMANATRABAJA', 'ESTU_TIPOREMUNERACION', 'COLE_COD_DANE_ESTABLECIMIENTO',
       'COLE_GENERO', 'COLE_NATURALEZA', 'COLE_CALENDARIO', 'COLE_BILINGUE',
       'COLE_CARACTER', 'COLE_AREA_UBICACION',
       'COLE_COD_MCPIO_UBICACION', 'COLE_MCPIO_UBICACION',
       'COLE_COD_DEPTO_UBICACION', 'COLE_DEPTO_UBICACION',
       'ESTU_COD_MCPIO_PRESENTACION', 'PUNT_GLOBAL',
                     'PERCENTIL_GLOBAL']]

#  Ver si hay valores nulos
data.isna().sum()

# Eliminar todos los valores nulos
data = data.dropna()

# Cuántas filas y columnas
data.shape

# Cuántos colegios, municipios y departamentos
count = data[['COLE_COD_DANE_ESTABLECIMIENTO','COLE_COD_MCPIO_UBICACION', 
              'COLE_COD_DEPTO_UBICACION' ]].nunique()
print(count)

# De qué tipo son las variables
data.dtypes

#Averiguar cuántos estudiantes de cada caracteristicas
count_internet = data.groupby(['FAMI_TIENEINTERNET']).FAMI_TIENEINTERNET.count()
count_comp = data.groupby(['FAMI_TIENECOMPUTADOR']).FAMI_TIENECOMPUTADOR.count()
count_etnia = data.groupby(['ESTU_TIENEETNIA']).ESTU_TIENEETNIA.count()
count_genero = data.groupby(['ESTU_GENERO']).ESTU_GENERO.count()
count_trabaja = data.groupby(['ESTU_HORASSEMANATRABAJA']).ESTU_HORASSEMANATRABAJA.count()

ind_variables = { "Tiene Internet": count_internet, "Tiene Computador": count_comp ,
             "Tiene Etnía": count_etnia, "Es Masculino": count_genero, "No Trabaja": count_trabaja}

print(ind_variables)

# Ver las diferentes categorías de las variables  
categoricas = data.loc[: ,["ESTU_GENERO"  ,  "ESTU_TIENEETNIA" , "FAMI_TIENEINTERNET" ,
            "FAMI_TIENECOMPUTADOR" ,"COLE_GENERO"  ,  "COLE_NATURALEZA" ,   "COLE_CALENDARIO" ,
            "COLE_BILINGUE" ,"COLE_CARACTER"  , "COLE_AREA_UBICACION", 'ESTU_HORASSEMANATRABAJA']]

for col in categoricas:
    print(data[col].unique())

# Crear dummy variables
data = pd.get_dummies(data, columns=["ESTU_GENERO"  ,  "ESTU_TIENEETNIA" , "FAMI_TIENEINTERNET" ,
            "FAMI_TIENECOMPUTADOR" ,"COLE_GENERO"  ,  "COLE_NATURALEZA" ,   "COLE_CALENDARIO" ,
            "COLE_BILINGUE" , "COLE_AREA_UBICACION", "ESTU_HORASSEMANATRABAJA", "COLE_CARACTER"])

# Ver los nombres
data.columns

#renombrar las columnas
data.rename(columns= {'ESTU_GENERO_M': 'male',
                      'ESTU_TIENEETNIA_Si' : 'etnia_si',
                      'FAMI_ESTRATOVIVIENDA_Estrato 2' : 'estrato_2', 
                      'FAMI_ESTRATOVIVIENDA_Estrato 3' : 'estrato_3',
                      'FAMI_ESTRATOVIVIENDA_Estrato 4' : 'estrato_4',
                      'FAMI_ESTRATOVIVIENDA_Estrato 5' : 'estrato_5',
                      'FAMI_ESTRATOVIVIENDA_Estrato 6' : 'estrato_6',
                      'FAMI_ESTRATOVIVIENDA_Sin Estrato' : 'estrato_sinestrato',
                      'FAMI_TIENEINTERNET_Si' : 'internet_si',
                      'FAMI_TIENECOMPUTADOR_Si': 'computador_si',
                      'COLE_GENERO_MASCULINO' : 'cole_masculino',
                      'COLE_NATURALEZA_NO OFICIAL' : 'cole_privado', 'COLE_NATURALEZA_OFICIAL' : 'cole_publico',
                      'COLE_CALENDARIO_A' : 'calendario_a',
                      'COLE_CALENDARIO_B' : 'calendario_b',
                      'COLE_CALENDARIO_OTRO' : 'calendario_otro',
                      'COLE_BILINGUE_S' : 'cole_bilingue', 
                      'COLE_BILINGUE_N' : 'cole_nbilingue',
                      'COLE_AREA_UBICACION_URBANO' : 'cole_urbano',
                      'COLE_AREA_UBICACION_RURAL' : 'cole_rural',
                      'ESTU_HORASSEMANATRABAJA_0' : 'trabaja_no',
                      'COLE_CARACTER_ACADÉMICO' : 'cole_academico'
                     }, inplace=True)

#Check
data.columns

data.to_csv("Datos/Datos.csv", index = False)
