==> 4. Regresiones.ipynb <==
import pandas as pd
import numpy as np
import statsmodels.api as sm
import statsmodels.formula.api as smf

df = pd.read_csv("Datos/Datos.csv")

df

df.columns

df.describe()

model = (smf.ols(formula= 'PUNT_GLOBAL ~ male + etnia_si + internet_si + computador_si + cole_academico + calendario_b + cole_bilingue + cole_urbano' ,data= df).fit())
model.params
print(model.summary())

df = pd.get_dummies(df, columns=['FAMI_ESTRATOVIVIENDA', 'FAMI_SITUACIONECONOMICA',
       'ESTU_DEDICACIONLECTURADIARIA', 'ESTU_DEDICACIONINTERNET', 'ESTU_TIPOREMUNERACION'])

df.columns

df.rename(columns= {'FAMI_ESTRATOVIVIENDA_Estrato 1': 'estrato1',
                      'FAMI_ESTRATOVIVIENDA_Estrato 2': 'estrato2',
                      'FAMI_ESTRATOVIVIENDA_Estrato 3': 'estrato3',
                      'FAMI_ESTRATOVIVIENDA_Estrato 4' : 'estrato4',
                      'FAMI_ESTRATOVIVIENDA_Estrato 5' : 'estrato5',
                      'FAMI_ESTRATOVIVIENDA_Estrato 6' : 'estrato6',
                      'FAMI_ESTRATOVIVIENDA_Sin Estrato' : 'estrato0',
                      'FAMI_SITUACIONECONOMICA_Igual' : 'econ_igual',
                      'FAMI_SITUACIONECONOMICA_Mejor': 'econ_mejor',
                      'FAMI_SITUACIONECONOMICA_Peor': 'econ_peor',
                      'ESTU_DEDICACIONLECTURADIARIA_30 minutos o menos' : 'lectura<30',
                      'ESTU_DEDICACIONLECTURADIARIA_Entre 1 y 2 horas' :'lectura1_2',
                      'ESTU_DEDICACIONLECTURADIARIA_Entre 30 y 60 minutos': 'lectura30_60',
                      'ESTU_DEDICACIONLECTURADIARIA_Más de 2 horas' : 'lectura>2',
                      'ESTU_DEDICACIONLECTURADIARIA_No leo por entretenimiento' : 'lectura_no',
                      'ESTU_DEDICACIONINTERNET_30 minutos o menos' : 'navega<30',
                      'ESTU_DEDICACIONINTERNET_Entre 1 y 3 horas' :'navega1_3',
                      'ESTU_DEDICACIONINTERNET_Entre 30 y 60 minutos' : 'navega30_60',
                      'ESTU_DEDICACIONINTERNET_Más de 3 horas' : 'navega>3',
                      'ESTU_DEDICACIONINTERNET_No Navega Internet' : 'navega_no',
                      'ESTU_TIPOREMUNERACION_No' : 'remuneracion_no',
                      'ESTU_TIPOREMUNERACION_Si, en efectivo' : 'remuneracion_si'}, inplace=True)

df.columns

model = (smf.ols(formula= 'PUNT_GLOBAL ~ male + trabaja_no + remuneracion_no + econ_peor + navega_no  + etnia_si +  + internet_si + computador_si + cole_publico + cole_academico + calendario_b + cole_bilingue + cole_urbano' ,data= df).fit())
model.params
print(model.summary())
